pyproject.toml
[project]
name = "cta-led-map-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==3.1.7",
    "dbt-core==1.11.6",
    "dbt-postgres==1.10.0",
    "pandas==3.0.1",
    "pgadmin4==9.12",
    "psycopg2-binary==2.9.11",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
]
